Unfortunately a few bits.  - Start of implementation for the VD geometry  - Rename the detectors directory to Detectors  - move the interface class to the Detectors directory
