+++ larpandora/LArPandoraInterface/Detectors/DUNEFarDetVDThreeView.h
@@ -1,4 +1,4 @@
-#include "larpandora/LArPandoraInterface/LArPandoraDetectorType.h"
+#include "larpandora/LArPandoraInterface/Detectors/LArPandoraDetectorType.h"
 
 #include "larcore/Geometry/Geometry.h"
 
@@ -6,13 +6,15 @@ namespace lar_pandora{
 
     class DUNEFarDetVDThreeView : public LArPandoraDetectorType {
     public:
-        geo::View_t TargetViewU() const override {return geo::kUnknown; };
+        DUNEFarDetVDThreeView();
+        geo::View_t TargetViewU() const override;
         geo::View_t TargetViewV() const override {return geo::kUnknown; };
         geo::View_t TargetViewW() const override {return geo::kUnknown; };
         float WirePitch(const geo::View_t view) const override {return 0.f; };
         float WireAngle(const geo::View_t view, const int tpc, const int cstat) const override {return 0.f; };
         bool ShouldSwitchUV(const unsigned int tpc, const unsigned int cstat) const override {return false; };
         void LoadDetectorGaps(LArDetectorGapList& listOfGaps) override {return; }; 
+    private:
+        //art::ServiceHandle<geo::Geometry> m_LArSoftGeometry;
     };
 }
-
