+++ larpandora/LArPandoraInterface/Detectors/LArPandoraDetectorType.h
